requirements.txt
ffmpeg-python
pymilvus
python-dotenv==1.0.1
streamlit==1.40
twelvelabs
